Release 2.0.0 (9th Jul 2019) 	License update: KVS SDK and GStreamer kvssink plugin are under Apache 2.0 license now. 	C-Producer - New feature 		Language level is C89. It should work for all GCC versions, tested minimum GCC version is 4.1. 		The Kinesis Video Streams C producer library wraps kinesis-video-pic (Platform Independent Code) with additional layer of API that allows scenario and platform-specific callbacks and events. 		Callbacks for sample retry logic implementation are now supported through C Producer 	CPP Producer - Updates 		Language level is C++11 instead of C++14. Required minimum GCC version is 4.8. 		Kinesis Video Producer CPP (kinesis-video-producer directory) is now revised to a wrapper of C Producer SDK which interfaces with the Platform Independent Repository (PIC, kinesis-video-pic directory). 		Platform Independent Code - New functionality 		Automatic CPD (codec private data) extraction from the stream 		Reset stream and reset connection functionality in PIC - In case of unrecoverable errors users can simply call reset stream; internally SDK will free & create the stream. 		Log level supported in PIC. It can be set in member variable of DeviceInfo.ClientInfo. 	Platform Independent Code - Bug Fixes 		Skip over error fragments - SDK will continue skip any invalid fragments are ingested through SDK earlier and continue streaming with valid ingested fragments. It will not just loop back and keep resending the broken fragment now. 		All unit tests (pic_test, cproducer_test, producer_test) are at C++11 level. Required minimum GCC version is 4.8.
	Please note the following changes if you have customized the SDK for your integration and not using samples directly.

		Changes to utility functions:
		RotatingStaticCredentialProvider is removed from C++ producer. Customer can use StaticCredentialProvider instead.
		CredentialProviderUtil and SerializedCredentials are removed from C++ producer.
		Following functions are removed from DefaultCallbackProvider: sleepUntilWithTimeCallback, shutdownStream, notifyResult, getControlPlaneUri, safeFreeBuffer, getStreamStatusFromString.
	Changes to the object model:
		Following attributes have been added: Version number is added to TrackInfo, Frame, StreamDescription, DeviceInfo, StreamInfo.
		Following attributes have been removed: certPath is removed from DeviceInfo in PIC.

+++ kinesis-video-pic/src/utils/tst/UtilTestFixture.h
@@ -100,7 +100,7 @@ public:
         // Validate the allocations cleanup
         DLOGI("Final remaining allocation size is %llu\n", gTotalUtilsMemoryUsage);
 
-        EXPECT_EQ(0, gTotalUtilsMemoryUsage);
+        EXPECT_EQ((UINT64) 0, gTotalUtilsMemoryUsage);
         globalMemAlloc = storedMemAlloc;
         globalMemAlignAlloc = storedMemAlignAlloc;
         globalMemCalloc = storedMemCalloc;
